requirements.txt
python-telegram-bot==20.7
aiohttp==3.8.6
python-dotenv==1.0.0